#1,2,3 Issues Addressed
I kept the initial design layout. However, I did change the colors,
color scheme in websites should be fairly static. This is because
disjointed colors leads to confusion and irritation .  This is the same
with changing text styles. Try to stick to one or two. Anymore than this
and it's noticeable to the viewer.

I also, fixed small HTML tabbing error. In addition, I got rid of the
extra CSS sheets as there was no discernible difference between
gutenberg.css and the others.